百度翻译spider/requirement.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.9
PyExecJS==1.5.1
requests==2.24.0
six==1.15.0
urllib3==1.25.9
